requirements-base.txt
distrax==0.1.3
optax==0.1.5
flax==0.6.11
numpy==1.22.4
black==23.3.0
pre-commit==3.3.3
argparse
wandb
tensorflow==2.13.0
tensorstore==0.1.40
orbax-checkpoint
pygame
gymnax
chex
matplotlib
imageio